added l+p tests

+++ src/main/resources/static/test/create.sql
@@ -1,5 +1,5 @@
 create table person (
-    id bigint,
+    id serial,
     firstName varchar(256), 
     lastName varchar(256), 
     address varchar(256),
